requirements.txt
djangorestframework
django
psycopg
django-extensions
typer
paramiko